+++ lib/plugins/aws/package/compile/layers.js
@@ -5,7 +5,7 @@ const BbPromise = require('bluebird');
 const _ = require('lodash');
 const path = require('path');
 const fsAsync = BbPromise.promisifyAll(require('fs'));
-const getLambdaLayerArtifactPath = require('../../../utils/getLambdaLayerArtifactPath');
+const getLambdaLayerArtifactPath = require('../../utils/getLambdaLayerArtifactPath');
 
 class AwsCompileLayers {
   constructor(serverless, options) {
