+++ lib/plugins/index.js
@@ -32,7 +32,7 @@ module.exports = [
   require('./aws/remove/index.js'),
   require('./aws/rollback/index.js'),
   require('./aws/rollbackFunction/index.js'),
-  require('./aws/package/compile/layers/index.js'),
+  require('./aws/package/compile/layers.js'),
   require('./aws/package/compile/functions.js'),
   require('./aws/package/compile/events/schedule.js'),
   require('./aws/package/compile/events/s3/index.js'),
